update make_protein_domain
this update includes a new strategy for pivot_wider. Without a unique ID to help identify groups, pivot_wider fails and makes list columns. So an easy way to prevent this is to make a col_id_helper column, that uses mutate + case_when to add a unique ID (row_number) when a new datagroup starts, and then fill to populate below. With this temp helper column, pivot wider knows which common IDs below to which groups.